(Removed unused App/AppTest classes)
Changed JBehave runner so that it'll fail on Pending implementations of steps (instead of ignoring).

Changed story to use first-person, e.g., "I create a new page..."

Wiki now tracks the concept of a "current page".

By default, the Wiki is created with a "Home" page that is the current page.